third_party/blink/web_tests/external/wpt/css/css-typed-om/stylevalue-normalization/normalize-image-expected.txt

This is a testharness.js-based test.
[FAIL] Normalizing a <gradient> returns a CSSImageValue
  assert_equals: expected "CSSImageValue" but got "CSSStyleValue"
Harness: the test ran to completion.
